README.md
This GitHub repository hosts the source code and documentation for eBlessings.live, an innovative platform designed to bridge Islamic teachings with modern technology. It serves as the central hub for the development, collaboration, and tracking of various projects under the eBlessings initiative. The repository is structured to facilitate easy access to code, educational content, AI-driven tools, and community-based projects.

Inside, you'll find:

Source Code: The backbone of our platform, including web applications, AI bots, and other software tools developed for eBlessings.live.
Documentation: Comprehensive guides, API documentation, and educational material that support our technology and educational offerings.
Project Management: Issue tracking, feature requests, and roadmaps that guide the development and improvement of our platform.
Collaborative Space: A place for developers, educators, and community members to contribute, discuss, and advance the goals of eBlessings.
By hosting our project on GitHub, we aim to foster transparency, collaboration, and community involvement in our journey to blend faith with technology for the betterment of the Ummah.